New jar + next sweet UI + drop indicator  + more
Added a new jar that is rounded, making this game more unique (REQUIRES UPSCALING TO HD). Also added a UI element that shows the player what the next sweet will be. Added a drop indicator line with an animation. Added a splash screen, main menu, and pause menu, all of which require designing, but are functional. Limited the x-position of the drop indicator to within the drop limits, above the jar's opening